// frontend/src/types/sale.ts
import { Seller } from "./seller";

export type SaleSum = {
    sellerName: string;
    sum: number;
}

export type SaleSucess = {
    sellerName: string;
    visited: number;
    deals: number;
}

export type Sale = {
    id: number;
    visited: number;
    deals: number;
    amount: number;
    date: string;
    seller: Seller;
}

export type Salepage = {
    content?: Sale[]
    last: boolean;
    totalElements: number;
    totalPages: number;
    size?: number;
    number: number;
    first: boolean;
    numberOfElements?: number;
    empty?: boolean;
}